ml-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
yfinance==0.2.40
xgboost==2.1.0
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
apscheduler==3.10.4
mysql-connector-python==9.0.0
